verify-test Create portal-verify-test-util and move common *TestCase classes to avoid package conflict
