+++ modules/apps/portal/portal-verify-test-util/bnd.bnd
@@ -1,0 +1,4 @@
+Bundle-Name: Liferay Portal Verify Test Utilities
+Bundle-SymbolicName: com.liferay.portal.verify.test.util
+Bundle-Version: 1.0.0
+Export-Package: com.liferay.portal.verify.test.util
