+++ modules/apps/portal/portal-verify-test-util/build.gradle
@@ -1,0 +1,11 @@
+dependencies {
+	compileOnly group: "com.liferay.portal", name: "com.liferay.portal.impl", version: "default"
+	compileOnly group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "default"
+	compileOnly group: "javax.portlet", name: "portlet-api", version: "3.0.0"
+	compileOnly group: "junit", name: "junit", version: "4.12"
+	compileOnly project(":core:registry-api")
+}
+
+liferay {
+	deployDir = file("${liferayHome}/osgi/test")
+}
