+++ modules/apps/portal/portal-verify-test-util/src/main/resources/com/liferay/portal/verify/test/util/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
